Add "lanterns" tags and initial dummy recipe
Craft a colored lantern with any glass pane and any other lantern. This
obviously needs to change when we have real colored lantern blocks /
items, but it's a start and proves that the custom tags are working
correctly.

Also tested the block tags in game using "execute if block <x> <y>
<z> #lanterncolors:lanterns"