Add advancements/recipes json files
Turns out you're supposed to directly extend the Minecraft
RecipeProvider rather than extending the ForgeRecipeProvider class. The
Forge one overrides the saveRecipeAdvancement method to do nothing, so
these json files were never generated and the custom recipes weren't
showing up in the in-game recipe book when you got the correct
materials.

Now we extend the correct class in our generator and the files are
properly created. This also adjusts the criterion so that you only need
to have a lantern to unlock the recipes and not glass panes.